+++ express_relay/examples/easy_lend/.gitignore
@@ -1,0 +1,5 @@
+lib/*
+out
+cache
+tslib
+!lib/README.md
