+++ express_relay/examples/easy_lend/README.md
@@ -1,0 +1,6 @@
+# EasyLend Protocol
+
+EasyLend is a simplified lending protocol that uses Express Relay for avoiding value leakage on liquidations.
+It uses Pyth price feeds to calculate the asset values and the liquidation thresholds.
+
+This project illustrates how to use the Express Relay SDK for contract integration and publishing opportunities.
