+++ express_relay/examples/easy_lend/contracts/EasyLendErrors.sol
@@ -1,0 +1,20 @@
+// Copyright (C) 2024 Lavra Holdings Limited - All Rights Reserved
+pragma solidity ^0.8.13;
+
+// Signature: 0xe922edfd
+error UncollateralizedVaultCreation();
+
+// Signature: 0xdcb430ee
+error InvalidVaultUpdate();
+
+// Signature: 0x9cd7b1c6
+error InvalidPriceExponent();
+
+// Signature: 0x85914873
+error InvalidLiquidation();
+
+// Signature: 0x61ca76d2
+error NegativePrice();
+
+// Signature: 0x4a7a3163
+error InvalidHealthRatios();
