+++ express_relay/examples/easy_lend/contracts/EasyLendStructs.sol
@@ -1,0 +1,13 @@
+// Copyright (C) 2024 Lavra Holdings Limited - All Rights Reserved
+pragma solidity ^0.8.13;
+
+struct Vault {
+    address tokenCollateral;
+    address tokenDebt;
+    uint256 amountCollateral;
+    uint256 amountDebt;
+    uint256 minHealthRatio; // 10**18 is 100%
+    uint256 minPermissionLessHealthRatio;
+    bytes32 tokenIdCollateral;
+    bytes32 tokenIdDebt;
+}
