+++ express_relay/examples/easy_lend/foundry.toml
@@ -1,0 +1,10 @@
+[profile.default]
+src = "contracts"
+out = "out"
+libs = [
+    'lib',
+    '../../../node_modules',
+    '../../../target_chains/ethereum/sdk/solidity'
+]
+
+# See more config options https://github.com/foundry-rs/foundry/blob/master/crates/config/README.md#all-options
