+++ express_relay/examples/easy_lend/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "easylend",
+  "version": "0.1.0",
+  "description": "Example lending protocol with express relay integration",
+  "private": true,
+  "files": [
+    "tslib/**/*"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "lint": "eslint src/",
+    "format": "prettier --write \"src/**/*.ts\"",
+    "monitor": "npm run build && node tslib/monitor.js",
+    "install-forge-deps": "forge install foundry-rs/forge-std@v1.7.6 --no-git --no-commit"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/pyth-network/pyth-crosschain.git"
+  },
+  "dependencies": {
+    "@openzeppelin/contracts": "^4.5.0",
+    "@pythnetwork/express-relay-evm-js": "*",
+    "@pythnetwork/express-relay-sdk-solidity": "*",
+    "@pythnetwork/pyth-evm-js": "*",
+    "@pythnetwork/pyth-sdk-solidity": "*",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.3.3",
+    "viem": "^2.7.6"
+  },
+  "devDependencies": {
+    "@types/yargs": "^17.0.10",
+    "eslint": "^8.56.0",
+    "prettier": "^2.6.2",
+    "typedoc": "^0.25.7",
+    "typescript": "^5.1",
+    "yargs": "^17.4.1"
+  }
+}
